requirements.txt
streamlit>=1.39.0
streamlit-aggrid
neo4j>=5.17.0
langchain>=0.3.7
langchain-openai>=0.0.2
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.0.0 